admin - securing the views

+++ products/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import render, redirect, reverse, get_object_or_404
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from django.db.models.functions import Lower
 from .models import Product, Category
@@ -71,8 +72,14 @@ def product_detail(request, product_id):
     return render(request, 'products/product_detail.html', context)
 
 
+@login_required
 def add_new_product(request):
     """ Add new product to AV Glasses store """
+
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owners can accsess this feature.')
+        return redirect(reverse('home'))
+
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES)
         if form.is_valid():
@@ -92,8 +99,15 @@ def add_new_product(request):
     return render(request, template, context)
 
 
+@login_required
 def edit_product(request, product_id):
     """ Edit a product in AV Glasses store """
+
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owners can accsess this feature.')
+        return redirect(reverse('home'))
+
+
     product = get_object_or_404(Product, pk=product_id)
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES, instance=product)
@@ -116,8 +130,15 @@ def edit_product(request, product_id):
     return render(request, template, context)
 
 
+@login_required
 def delete_product(request, product_id):
     """ Delete product from the AV Glasses store """
+
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owners can accsess this feature.')
+        return redirect(reverse('home'))
+
+
     product = get_object_or_404(Product, pk=product_id)
     product.delete()
     messages.success(request, 'Product successfuly deleted!')
